configs/imx8mn_bsh_smm_s2_defconfig: new defconfig
BSH SMM S2 (SystemMaster S2) Board is an add-on board which provide
input and output interfaces to a dedicated carrier board. It is designed
mainly to provide graphical/video and connectivity interfaces to the
appliance.

Board support package includes the following components:
- mainline Linux kernel 5.18
- mainline U-Boot 2022-07-rc2
- default packages from buildroot

Patches applied to the Linux kernel and U-Boot have been submitted but
not merged yet.

Co-authored-by: Michael Trimarchi <michael@amarulasolutions.com>
Signed-off-by: Michael Trimarchi <michael@amarulasolutions.com>
Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>